requirements.txt
flask==2.0.1
openai>=1.0.0
python-dotenv==0.19.0
yt-dlp==2023.11.16
requests>=2.31.0
